chojungho/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.poetry]
authors = ["jhcho <cho418651@ssrinc.co.kr>"]
description = ""
name = "url-shortcut"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
asyncpg = "^0.30.0"
black = "^24.10.0"
fastapi = {extras = ["standard"], version = "^0.115.3"}
flake8 = "^7.1.1"
isort = "^5.13.2"
pre-commit = "^4.0.1"
psycopg2 = "^2.9.10"
pybase62 = "^1.0.0"
python = "3.12.7"
redis = "^5.2.0"
sqlmodel = "^0.0.22"
toml-sort = "^0.23.1"
